the SuperShader API has been significantly streamlined.
The reason for this change was to make it easier to write your own custom SuperShader passes, e.g., to add new types of light sources like caustics and spotlights, or to create a deferred rendering system.

There is now a class SuperShader::Pass that can be subclassed to create a customized SuperShader. PosedModel will shortly have a method that supports rendering these custom passes.