I changed the layout algorithm for GuiControl captions.  If the caption is "", the control now fills its entire rect.  If the caption is " " or any other non-empty string, the control reserves space for its caption.  You can control the amount of space reserved by calling setCaptionSize.
This fixed the GuiNumberBox layout bugs and also made it much easier to create customized Guis (I always felt like I was fighting the captions before when trying to do anything other than the default layout).

I rewrote CameraControlWindow to compensate for the new layout. The constructor got slightly shorter and needs one fewer member variable, which is probably a sign that this was a good change.